requirements.txt
django
gunicorn
uvicorn
pytest
djangorestframework